Port to fabric

+++ src/main/java/com/rcx/materialis/util/MaterialisByproduct.java
@@ -40,7 +40,7 @@ public enum MaterialisByproduct implements IByproduct {
 	private final boolean alwaysPresent;
 	private final Supplier<? extends Fluid> fluidSupplier;
 	@Getter
-	private final int amount;
+	private final long amount;
 
 	MaterialisByproduct(boolean alwaysPresent, Supplier<? extends Fluid> fluidSupplier) {
 		this.name = name().toLowerCase(Locale.ROOT);
